+++ PlayBuilder/Services/IAtlasComparisonService.cs
@@ -1,0 +1,8 @@
+using PlayBuilder.Models;
+
+namespace PlayBuilder.Services;
+
+public interface IAtlasComparisonService
+{
+    AtlasComparisonReport Compare(ArchiveScanResult scan, CollectionRuleOptions options);
+}